+++ Screens/NavDetail.js
@@ -1,6 +1,5 @@
 //----------REACT UTILS-----------
 import React, { useState } from "react";
-
 //
 //
 //----------REDUX UTILS-----------
@@ -21,7 +20,6 @@ import Btn from "./Helpers/Btns.js";
 //
 //-------STYLES-------
 import globalStyles from "./GlobalStyles.js";
-
 //
 //
 //-------INITIALIZATIONS-------
@@ -63,11 +61,13 @@ export default function NavDetail({ navigation }) {
           <Text style={styles.title}>Resto Book</Text>
         </View>
         <View style={styles.btnMenu}>
-          <Btn
-            nombre="Add Food!"
-            ruta="AddMenuResto"
-            navigation={navigation}
-          />
+          {owner &&
+            <Btn
+              nombre="Add Food!"
+              ruta="AddMenuResto"
+              navigation={navigation}
+            />
+          }
         </View>
       </View>
     </View>
